simplify sentences converter , move split_into_words to TxtHelper.
